Fix bug with room duplication
There was no mechanism to keep track of who was already in a private
room with one another so one user can repeatedly open new private rooms
with the same user. To resolve this, a list keeping track of which
users the current user has a private room with then disabling the
ability to start a new room with them.