Fix messages bug and add default room label
Messages were not being sent because the dispatched object did not have
the expected payload shape. Messages were also not updating on the
sender's chat because the socket server broadcasts the message to
all other connected clients except the sender, therefore the message
must be added to the state manually.